@@ -1,24 +1,20 @@
 {
   "name": "react-component-update",
-  "version": "1.5.0",
+  "version": "1.6.0",
   "description": "Extends the native React Component to streamline updates",
   "main": "lib/index.js",
   "files": [
     "lib/**/!(*.spec).js"
   ],
   "engines": {
-    "node": ">=10"
+    "node": ">=12"
   },
   "scripts": {
-    "clean": "rimraf lib",
-    "build": "npm run clean && babel src --out-dir lib",
-    "build-and-test": "npm run test && npm run build",
-    "install-react": "enzyme-adapter-react-install 16",
+    "build": "babel src --out-dir lib --delete-dir-on-start",
+    "changelog": "npx auto-changelog",
     "lint": "eslint ./ --report-unused-disable-directives",
-    "test": "jest",
-    "travisci": "npm run install-react && npm run lint && npm run test && codecov",
-    "prepublish": "npx publish-please guard",
-    "publish-please": "npx publish-please"
+    "release": "npx release-it --no-increment",
+    "test": "jest"
   },
   "repository": {
     "type": "git",
@@ -35,43 +31,23 @@
     "@babel/core": "^7.2.2",
     "@babel/preset-env": "^7.2.0",
     "@babel/preset-react": "^7.0.0",
-    "babel-jest": "^26.3.0",
-    "chai": "^4.0.2",
-    "codecov": "^3.0.0",
+    "babel-jest": "^27.0.1",
     "create-react-class": "^15.6.0",
-    "dirty-chai": "^2.0.0",
     "enzyme": "^3.6.0",
     "enzyme-adapter-react-16": "^1.5.0",
     "enzyme-adapter-react-helper": "^1.3.0",
-    "eslint": "^7.8.0",
-    "eslint-config-airbnb": "^18.2.0",
-    "eslint-plugin-import": "^2.13.0",
+    "eslint": "^8.4.1",
+    "eslint-config-airbnb": "^19.0.2",
+    "eslint-plugin-import": "^2.25.3",
     "eslint-plugin-jsx-a11y": "^6.1.0",
-    "eslint-plugin-react": "^7.11.0",
-    "jest": "^26.4.2",
-    "lodash.uniqueid": "^4.0.1",
-    "react": "*",
-    "react-dom": "*",
-    "react-test-renderer": "*",
-    "rimraf": "^3.0.2",
-    "sinon": "^9.0.3",
-    "sinon-chai": "^3.0.0"
+    "eslint-plugin-react": "^7.27.1",
+    "jest": "^27.0.3",
+    "jest-extended": "^1.2.0",
+    "react": "<17",
+    "react-dom": "<17",
+    "react-test-renderer": "<17"
   },
   "peerDependencies": {
-    "react": "*"
-  },
-  "jest": {
-    "coverageDirectory": "./coverage/",
-    "collectCoverage": true,
-    "collectCoverageFrom": [
-      "src/**/!(*.spec).js"
-    ],
-    "setupFiles": [
-      "<rootDir>/test-shims.js",
-      "<rootDir>/test-setup.js"
-    ],
-    "testMatch": [
-      "**/src/*.spec.js?(x)"
-    ]
+    "react": "<17"
   }
 }
